fix: shadow style
