+++ packages/mask/src/plugins/Avatar/SNSAdaptor/RainbowBox.tsx
@@ -2,30 +2,32 @@ import { keyframes, makeStyles, useStylesExtends } from '@masknet/theme'
 import type { Keyframes } from '@emotion/serialize'
 
 export const rainbowBorderKeyFrames: Keyframes = keyframes`
-0%,to {
-    border-color: #00f8ff;
-    box-shadow: 0 5px 15px rgba(0,248,255,.4),0 10px 30px rgba(37,41,46,.2);
-}
-
-20% {
-    border-color: #a4ff00;
-    box-shadow: 0 5px 15px rgba(164,255,0,.4),0 10px 30px rgba(37,41,46,.2);
-}
-
-40% {
-    border-color: #f7275e;
-    box-shadow: 0 5px 15px rgba(247,39,94,.4),0 10px 30px rgba(37,41,46,.2)
-}
-
-60% {
-    border-color: #ffd300;
-    box-shadow: 0 5px 15px rgba(255,211,0,.4),0 10px 30px rgba(37,41,46,.2)
-}
-
-80% {
-    border-color: #ff8a00;
-    box-shadow: 0 5px 15px rgba(255,138,0,.4),0 10px 30px rgba(37,41,46,.2)
-}
+    0%,
+    to {
+        border-color: #00f8ff;
+        -webkit-box-shadow: 0 5px 15px rgba(0, 248, 255, 0.4), 0 10px 30px rgba(37, 41, 46, 0.2);
+        box-shadow: 0 5px 15px rgba(0, 248, 255, 0.4), 0 10px 30px rgba(37, 41, 46, 0.2);
+    }
+    20% {
+        border-color: #a4ff00;
+        -webkit-box-shadow: 0 5px 15px rgba(164, 255, 0, 0.4), 0 10px 30px rgba(37, 41, 46, 0.2);
+        box-shadow: 0 5px 15px rgba(164, 255, 0, 0.4), 0 10px 30px rgba(37, 41, 46, 0.2);
+    }
+    40% {
+        border-color: #f7275e;
+        -webkit-box-shadow: 0 5px 15px rgba(247, 39, 94, 0.4), 0 10px 30px rgba(37, 41, 46, 0.2);
+        box-shadow: 0 5px 15px rgba(247, 39, 94, 0.4), 0 10px 30px rgba(37, 41, 46, 0.2);
+    }
+    60% {
+        border-color: #ffd300;
+        -webkit-box-shadow: 0 5px 15px rgba(255, 211, 0, 0.4), 0 10px 30px rgba(37, 41, 46, 0.2);
+        box-shadow: 0 5px 15px rgba(255, 211, 0, 0.4), 0 10px 30px rgba(37, 41, 46, 0.2);
+    }
+    80% {
+        border-color: #ff8a00;
+        -webkit-box-shadow: 0 5px 15px rgba(255, 138, 0, 0.4), 0 10px 30px rgba(37, 41, 46, 0.2);
+        box-shadow: 0 5px 15px rgba(255, 138, 0, 0.4), 0 10px 30px rgba(37, 41, 46, 0.2);
+    }
 `
 
 interface StyleProps {
@@ -38,7 +40,7 @@ const useStyles = makeStyles<StyleProps>()((theme, { width, height, radius = '10
         animation: `${rainbowBorderKeyFrames} 6s linear infinite`,
         width,
         height,
-        boxShadow: '0 5px 15px rgba(0,248,255,.4),0 10px 30px rgba(37,41,46,.2)',
+        boxShadow: '0 5px 15px rgba(0, 248, 255, 0.4), 0 10px 30px rgba(37, 41, 46, 0.2)',
         transition: '.125s ease',
         borderRadius: radius,
         display: 'flex',
